+++ examples/04_Scenarios/scenario_example.py
@@ -123,7 +123,7 @@ if __name__ == '__main__':
             size=50,
             relative_minimum=0.1,
             relative_maximum=1,
-            on_off_parameters=fx.OnOffParameters(),
+            status_parameters=fx.StatusParameters(),
         ),
         fuel_flow=fx.Flow(label='Q_fu', bus='Gas'),
     )
@@ -135,7 +135,7 @@ if __name__ == '__main__':
         thermal_efficiency=0.48,  # Realistic thermal efficiency (48%)
         electrical_efficiency=0.40,  # Realistic electrical efficiency (40%)
         electrical_flow=fx.Flow(
-            'P_el', bus='Strom', size=60, relative_minimum=5 / 60, on_off_parameters=fx.OnOffParameters()
+            'P_el', bus='Strom', size=60, relative_minimum=5 / 60, status_parameters=fx.StatusParameters()
         ),
         thermal_flow=fx.Flow('Q_th', bus='Fernwärme'),
         fuel_flow=fx.Flow('Q_fu', bus='Gas'),